Massive architectural refactor
Previously, schema construction was divided into two phases:
Mapping and Ref resolution. Going forward, this would have required us
to awkwardly split some mapping logic, as some of it depends on Refs to
other components being resolved.

Instead, we use the fact that only top-level components can be addressed
by name: First, all top-level elements are scanned and their name as
well as the element itself are recorded; additionally, Refs for the
component are pre-allocated (without any content).
This can then be used to potentially process these nodes out-of-order:
When one component needs data from another component which has not yet
been processed, it requests the associated Ref from the Mapping Context
which then processes the required node.
To prevent cycles, the list of in-progress nodes is stored in the
context.