Implement complex content type for extension derivation
The spec is not particularly helpful in regards to saying what _can_
be absent or which kind of term a {term} property is at specific points
where it expects a property only available on model groups.

Hence, this code needs a thorough inspection on what assumptions can
safely be made.