Move ownership of the shared context to the caller
Previously, the "core" context was owned by the root MappingContext,
and every child context had a reference to this exact core context.
This meant that every MappingContext needed a Cow-like member
to decide whether it owned the root context or not.

With this commit, we instead let the caller of Schema::map_from_xml()
create the RootContext, and only create the MappingContext locally in
the schema mapper.

This approach still needs some code duplication, but these are mostly
accessors for each context's members.