Reorganize the core model and the generator into separate crates
This commit introduces a more logical split between the core model and
the generator, which previously were modules in the same crate. With
this change, the xsd model/parser could now be used on its own as a
library.

Some cleanup is required to remove artifacts like duplicate option
enums.