Make components keep track of whether they are built-in
Instead of checking if the name is the one of a built-in, relevant
components (Simple/ComplexTypeDef, AttributeDecl) now store an
is_builtin property which is only true if the component was created by
the builtin module.

This approach should be more robust, since double-overwriting builtins
should not be possible anyway, and it is easier for the generator to
decide whether to generate a type or use an existing primitive type
(like String) based on this.