Add parsing implementation of complex type with simple content
This also changes how rust-builtin primitive types are handled; instead
of using the actual name (like String) to define a struct in the
builtins module, it now uses a generic struct `PrimitiveType<T>` that
impls e.g. `SimpleType<String>` according to the spec, and provides an
`into_inner()`. The old implementation will be removed later, since that
breaks compatibility for any users like the test-runner.

It's also increasingly clear that a bigger refactor of the generator is
needed, since it's growing unmaintainable, but for now let's push
forward until we have a semi-working solution.